*修改了StaticStringsDict的储存方法，需要两倍的Bucket储存空间 *DWORD定义由unsigned long改为u32 *DataPool中成员变量和Variable大部分情况下，查找的时间复杂度由O(ln(n))降为O(2)。

+++ clstd/cltypes.h
@@ -119,8 +119,8 @@ typedef wchar_t             wch;
 
 typedef u8                CLBYTE;
 typedef u8*               CLLPBYTE;
-typedef unsigned short    CLWORD;
-typedef unsigned long     CLDWORD;
+typedef u16               CLWORD;
+typedef u32               CLDWORD;
 typedef u32               CLUINT;
 typedef u32*              CLLPUINT;
 typedef u32               CLULONG;
